+++ Calculator/CalculatorProjectTests/CalculatorProjectTests.swift
@@ -4,7 +4,8 @@ import XCTest
 @testable import Calculator
 
 class CalculatorItemQueueTests: XCTestCase {
-    var sut: CalculatorItemQueue<String>!
+    var sut: CalculatorItemQueue<T: CalculateItemProtocol>!
+    
     
     override func setUpWithError() throws {
         try super.setUpWithError()
